Move Redux logic
